students join classes
This is what it should take for students to join a class however for some reason it it not fucking working. I have no clue what is up with this shit. PHP checker finds nothing and i can track it back to line 47 but it some how still wont work no mater what i do.