Fixes (untested) 8/23
These should be the fixes nessarry to have the class page be fully functional aswell as directing people to the correct dashboard and creating a slot in the assign space for them when they sign up. Remind me to filter out teachers from the assign database however it might be a good thing to have teachers be like super students.